Switch to Vite
