+++ package.json
@@ -7,25 +7,25 @@
   "displayName": "RFC 3339 vs. ISO 8601",
   "screenshot": "/res/Screenshot.png",
   "dependencies": {
+    "@babel/preset-env": "^7.22.15",
+    "@babel/preset-react": "^7.22.15",
     "@babel/register": "^7.17.7",
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
+    "@vitejs/plugin-react": "^4.0.4",
     "babel-plugin-transform-assets": "^1.0.2",
     "convert-svg-to-png": "^0.6.2",
     "gh-pages": "^3.2.3",
     "iana-timezones": "1.1.2022-e",
-    "react": "^18.1.0",
-    "react-dom": "^18.1.0",
-    "react-scripts": "^5.0.1",
-    "web-vitals": "^1.0.1"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "vite": "^4.4.9"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "deploy": "gh-pages -d build",
+    "start": "vite",
+    "build": "vite build",
+    "deploy": "gh-pages -d dist",
     "readme": "node ./scripts/generateReadme.js"
   },
   "eslintConfig": {
