+++ src/index.jsx
@@ -3,6 +3,7 @@ import { createRoot } from 'react-dom/client';
 import './index.css';
 import App from './App';
 
+// @ts-ignore
 const root = createRoot(document.getElementById('root'));
 
 root.render(
